+++ x/auth/types/types.proto
@@ -18,6 +18,16 @@ message BaseAccount {
   uint64 sequence       = 4;
 }
 
+// ModuleAccount defines an account for modules that holds coins on a pool
+message ModuleAccount {
+  option (gogoproto.goproto_getters)  = false;
+  option (gogoproto.goproto_stringer) = false;
+
+  BaseAccount     base_account = 1 [(gogoproto.embed) = true, (gogoproto.moretags) = "yaml:\"base_account\""];
+  string          name         = 2;
+  repeated string permissions  = 3;
+}
+
 // Params defines the parameters for the auth module.
 message Params {
   option (gogoproto.equal)            = true;
@@ -27,8 +37,7 @@ message Params {
   uint64 tx_sig_limit            = 2 [(gogoproto.moretags) = "yaml:\"tx_sig_limit\""];
   uint64 tx_size_cost_per_byte   = 3 [(gogoproto.moretags) = "yaml:\"tx_size_cost_per_byte\""];
   uint64 sig_verify_cost_ed25519 = 4
-  [(gogoproto.customname) = "SigVerifyCostED25519", (gogoproto.moretags) = "yaml:\"sig_verify_cost_ed25519\""];
+      [(gogoproto.customname) = "SigVerifyCostED25519", (gogoproto.moretags) = "yaml:\"sig_verify_cost_ed25519\""];
   uint64 sig_verify_cost_secp256k1 = 5
-  [(gogoproto.customname) = "SigVerifyCostSecp256k1", (gogoproto.moretags) = "yaml:\"sig_verify_cost_secp256k1\""];
+      [(gogoproto.customname) = "SigVerifyCostSecp256k1", (gogoproto.moretags) = "yaml:\"sig_verify_cost_secp256k1\""];
 }
-
