+++ x/bank/exported/exported.go
@@ -2,7 +2,6 @@ package exported
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	authexported "github.com/cosmos/cosmos-sdk/x/auth/exported"
 )
 
 // GenesisBalance defines a genesis balance interface that allows for account
@@ -10,16 +9,6 @@ import (
 type GenesisBalance interface {
 	GetAddress() sdk.AccAddress
 	GetCoins() sdk.Coins
-}
-
-// ModuleAccountI defines an account interface for modules that hold tokens in
-// an escrow.
-type ModuleAccountI interface {
-	authexported.Account
-
-	GetName() string
-	GetPermissions() []string
-	HasPermission(string) bool
 }
 
 // SupplyI defines an inflationary supply interface for modules that handle
