+++ x/bank/types/types.proto
@@ -3,7 +3,6 @@ package cosmos_sdk.x.bank.v1;
 
 import "third_party/proto/gogoproto/gogo.proto";
 import "types/types.proto";
-import "x/auth/types/types.proto";
 
 option go_package = "github.com/cosmos/cosmos-sdk/x/bank/types";
 
@@ -47,17 +46,6 @@ message MsgMultiSend {
 
   repeated Input  inputs  = 1 [(gogoproto.nullable) = false];
   repeated Output outputs = 2 [(gogoproto.nullable) = false];
-}
-
-// ModuleAccount defines an account for modules that holds coins on a pool
-message ModuleAccount {
-  option (gogoproto.goproto_getters)  = false;
-  option (gogoproto.goproto_stringer) = false;
-
-  cosmos_sdk.x.auth.v1.BaseAccount base_account = 1
-      [(gogoproto.embed) = true, (gogoproto.moretags) = "yaml:\"base_account\""];
-  string          name        = 2;
-  repeated string permissions = 3;
 }
 
 // Supply represents a struct that passively keeps track of the total supply
